move LLMRequestBody from scheduling to requesthandling
