+++ pkg/epp/framework/plugins/requesthandling/parsers/openai/openai.go
@@ -23,9 +23,7 @@ import (
 	"strings"
 
 	fwkplugin "sigs.k8s.io/gateway-api-inference-extension/pkg/epp/framework/interface/plugin"
-	fwkrc "sigs.k8s.io/gateway-api-inference-extension/pkg/epp/framework/interface/requestcontrol"
 	fwkrh "sigs.k8s.io/gateway-api-inference-extension/pkg/epp/framework/interface/requesthandling"
-	"sigs.k8s.io/gateway-api-inference-extension/pkg/epp/framework/interface/scheduling"
 )
 
 const (
@@ -86,7 +84,7 @@ func (p *OpenAIParser) WithName(name string) *OpenAIParser {
 }
 
 // ParseRequest parses the request body and headers and returns a map representation.
-func (p *OpenAIParser) ParseRequest(ctx context.Context, body []byte, headers map[string]string) (*scheduling.LLMRequestBody, error) {
+func (p *OpenAIParser) ParseRequest(ctx context.Context, body []byte, headers map[string]string) (*fwkrh.LLMRequestBody, error) {
 	bodyMap := make(map[string]any)
 	if err := json.Unmarshal(body, &bodyMap); err != nil {
 		return nil, errors.New("error unmarshaling request bodyMap")
@@ -175,39 +173,39 @@ func determineAPITypeFromPath(path string) string {
 }
 
 // extractRequestBody extracts the LLMRequestBody from the given request body map using path-based detection.
-func extractRequestBody(rawBody []byte, headers map[string]string) (*scheduling.LLMRequestBody, error) {
+func extractRequestBody(rawBody []byte, headers map[string]string) (*fwkrh.LLMRequestBody, error) {
 	// Determine API type from request path
 	path := getRequestPath(headers)
 	apiType := determineAPITypeFromPath(path)
 
 	switch apiType {
 	case conversationsAPI:
-		var conversations scheduling.ConversationsRequest
+		var conversations fwkrh.ConversationsRequest
 		if err := json.Unmarshal(rawBody, &conversations); err == nil && len(conversations.Items) > 0 {
-			return &scheduling.LLMRequestBody{Conversations: &conversations}, nil
+			return &fwkrh.LLMRequestBody{Conversations: &conversations}, nil
 		}
 		return nil, errors.New("invalid conversations request: must have items field")
 
 	case responsesAPI:
-		var responses scheduling.ResponsesRequest
+		var responses fwkrh.ResponsesRequest
 		if err := json.Unmarshal(rawBody, &responses); err == nil && responses.Input != nil {
-			return &scheduling.LLMRequestBody{Responses: &responses}, nil
+			return &fwkrh.LLMRequestBody{Responses: &responses}, nil
 		}
 		return nil, errors.New("invalid responses request: must have input field")
 
 	case chatCompletionsAPI:
-		var chatCompletions scheduling.ChatCompletionsRequest
+		var chatCompletions fwkrh.ChatCompletionsRequest
 		if err := json.Unmarshal(rawBody, &chatCompletions); err == nil {
 			if err = validateChatCompletionsMessages(chatCompletions.Messages); err == nil {
-				return &scheduling.LLMRequestBody{ChatCompletions: &chatCompletions}, nil
+				return &fwkrh.LLMRequestBody{ChatCompletions: &chatCompletions}, nil
 			}
 		}
 		return nil, errors.New("invalid chat completions request: must have valid messages field")
 
 	case completionsAPI:
-		var completions scheduling.CompletionsRequest
+		var completions fwkrh.CompletionsRequest
 		if err := json.Unmarshal(rawBody, &completions); err == nil && completions.Prompt != "" {
-			return &scheduling.LLMRequestBody{Completions: &completions}, nil
+			return &fwkrh.LLMRequestBody{Completions: &completions}, nil
 		}
 		return nil, errors.New("invalid completions request: must have prompt field")
 
@@ -216,14 +214,14 @@ func extractRequestBody(rawBody []byte, headers map[string]string) (*scheduling.
 	}
 }
 
-func validateChatCompletionsMessages(messages []scheduling.Message) error {
+func validateChatCompletionsMessages(messages []fwkrh.Message) error {
 	if len(messages) == 0 {
 		return errors.New("chat-completions request must have at least one message")
 	}
 	return nil
 }
 
-func extractUsage(responseBytes []byte) (*fwkrc.Usage, error) {
+func extractUsage(responseBytes []byte) (*fwkrh.Usage, error) {
 	var responseErr error
 	var responseBody map[string]any
 	responseErr = json.Unmarshal(responseBytes, &responseBody)
@@ -238,7 +236,7 @@ func extractUsage(responseBytes []byte) (*fwkrc.Usage, error) {
 		if usg["prompt_token_details"] != nil {
 			detailsMap := usg["prompt_token_details"].(map[string]any)
 			if cachedTokens, ok := detailsMap["cached_tokens"]; ok {
-				usage.PromptTokenDetails = &fwkrc.PromptTokenDetails{
+				usage.PromptTokenDetails = &fwkrh.PromptTokenDetails{
 					CachedTokens: int(cachedTokens.(float64)),
 				}
 			}
@@ -250,8 +248,8 @@ func extractUsage(responseBytes []byte) (*fwkrc.Usage, error) {
 
 // extractUsageByAPIType extracts usage statistics using the appropriate field names
 // based on the OpenAI API type identified by the "object" field.
-func extractUsageByAPIType(usg map[string]any, objectType string) fwkrc.Usage {
-	usage := fwkrc.Usage{}
+func extractUsageByAPIType(usg map[string]any, objectType string) fwkrh.Usage {
+	usage := fwkrh.Usage{}
 
 	switch {
 	case strings.HasPrefix(objectType, objectTypeResponse) || strings.HasPrefix(objectType, objectTypeConversation):
@@ -311,10 +309,10 @@ func extractUsageByAPIType(usg map[string]any, objectType string) fwkrc.Usage {
 //	data: {"response":{"usage":{"input_tokens":31,..},...},"type":"response.completed"}
 //
 // It extracts usage from events with type="response.completed".
-func extractUsageStreaming(responseText string) *fwkrc.Usage {
+func extractUsageStreaming(responseText string) *fwkrh.Usage {
 
 	var streamResponse struct {
-		Usage    *fwkrc.Usage `json:"usage"`
+		Usage    *fwkrh.Usage `json:"usage"`
 		Response struct {
 			Usage  map[string]any `json:"usage"`
 			Object string         `json:"object"`
