+++ pkg/epp/framework/plugins/requesthandling/parsers/openai/openai_test.go
@@ -22,10 +22,8 @@ import (
 	"testing"
 
 	"github.com/google/go-cmp/cmp"
-	"sigs.k8s.io/gateway-api-inference-extension/pkg/epp/framework/interface/scheduling"
 
 	fwkplugin "sigs.k8s.io/gateway-api-inference-extension/pkg/epp/framework/interface/plugin"
-	fwkrc "sigs.k8s.io/gateway-api-inference-extension/pkg/epp/framework/interface/requestcontrol"
 	fwkrh "sigs.k8s.io/gateway-api-inference-extension/pkg/epp/framework/interface/requesthandling"
 )
 
@@ -49,7 +47,7 @@ func TestOpenAIParser_ParseRequest(t *testing.T) {
 		name    string
 		headers map[string]string
 		body    map[string]any
-		want    *scheduling.LLMRequestBody
+		want    *fwkrh.LLMRequestBody
 		wantErr bool
 	}{
 		{
@@ -59,8 +57,8 @@ func TestOpenAIParser_ParseRequest(t *testing.T) {
 				"model":  "test",
 				"prompt": "test prompt",
 			},
-			want: &scheduling.LLMRequestBody{
-				Completions: &scheduling.CompletionsRequest{
+			want: &fwkrh.LLMRequestBody{
+				Completions: &fwkrh.CompletionsRequest{
 					Prompt: "test prompt",
 				},
 				ParsedBody: map[string]any{
@@ -83,11 +81,11 @@ func TestOpenAIParser_ParseRequest(t *testing.T) {
 					},
 				},
 			},
-			want: &scheduling.LLMRequestBody{
-				ChatCompletions: &scheduling.ChatCompletionsRequest{
-					Messages: []scheduling.Message{
-						{Role: "system", Content: scheduling.Content{Raw: "this is a system message"}},
-						{Role: "user", Content: scheduling.Content{Raw: "hello"}},
+			want: &fwkrh.LLMRequestBody{
+				ChatCompletions: &fwkrh.ChatCompletionsRequest{
+					Messages: []fwkrh.Message{
+						{Role: "system", Content: 	fwkrh.Content{Raw: "this is a system message"}},
+						{Role: "user", Content: fwkrh.Content{Raw: "hello"}},
 					},
 				},
 				ParsedBody: map[string]any{
@@ -131,22 +129,22 @@ func TestOpenAIParser_ParseRequest(t *testing.T) {
 					},
 				},
 			},
-			want: &scheduling.LLMRequestBody{
-				ChatCompletions: &scheduling.ChatCompletionsRequest{
-					Messages: []scheduling.Message{
-						{Role: "system", Content: scheduling.Content{
-							Structured: []scheduling.ContentBlock{
+			want: &fwkrh.LLMRequestBody{
+				ChatCompletions: &fwkrh.ChatCompletionsRequest{
+					Messages: []fwkrh.Message{
+						{Role: "system", Content: fwkrh.Content{
+							Structured: []fwkrh.ContentBlock{
 								{
 									Text: "Describe this image in one sentence.",
 									Type: "text",
 								},
 							},
 						}},
-						{Role: "user", Content: scheduling.Content{
-							Structured: []scheduling.ContentBlock{
+						{Role: "user", Content: fwkrh.Content{
+							Structured: []fwkrh.ContentBlock{
 								{
 									Type:     "image_url",
-									ImageURL: scheduling.ImageBlock{Url: "https://example.com/images/dui.jpg."},
+									ImageURL: fwkrh.ImageBlock{Url: "https://example.com/images/dui.jpg."},
 								},
 							},
 						}},
@@ -204,18 +202,18 @@ func TestOpenAIParser_ParseRequest(t *testing.T) {
 					},
 				},
 			},
-			want: &scheduling.LLMRequestBody{
-				ChatCompletions: &scheduling.ChatCompletionsRequest{
-					Messages: []scheduling.Message{
-						{Role: "user", Content: scheduling.Content{
-							Structured: []scheduling.ContentBlock{
+			want: &fwkrh.LLMRequestBody{
+				ChatCompletions: &fwkrh.ChatCompletionsRequest{
+					Messages: []fwkrh.Message{
+						{Role: "user", Content: fwkrh.Content{
+							Structured: []fwkrh.ContentBlock{
 								{
 									Type:       "input_audio",
-									InputAudio: scheduling.AudioBlock{Data: "base64data", Format: "wav"},
+									InputAudio: fwkrh.AudioBlock{Data: "base64data", Format: "wav"},
 								},
 								{
 									Type:     "video_url",
-									VideoURL: scheduling.VideoBlock{Url: "https://example.com/video.mp4"},
+									VideoURL: fwkrh.VideoBlock{Url: "https://example.com/video.mp4"},
 								},
 							},
 						}},
@@ -262,9 +260,9 @@ func TestOpenAIParser_ParseRequest(t *testing.T) {
 				"add_generation_prompt":        true,
 				"chat_template_kwargs":         map[string]any{"key": "value"},
 			},
-			want: &scheduling.LLMRequestBody{
-				ChatCompletions: &scheduling.ChatCompletionsRequest{
-					Messages:                  []scheduling.Message{{Role: "user", Content: scheduling.Content{Raw: "hello"}}},
+			want: &fwkrh.LLMRequestBody{
+				ChatCompletions: &fwkrh.ChatCompletionsRequest{
+					Messages:                  []fwkrh.Message{{Role: "user", Content: fwkrh.Content{Raw: "hello"}}},
 					Tools:                     []any{map[string]any{"type": "function"}},
 					Documents:                 []any{map[string]any{"content": "doc"}},
 					ChatTemplate:              "custom template",
@@ -443,8 +441,8 @@ func TestOpenAIParser_ParseRequest(t *testing.T) {
 				"prompt":     "test prompt",
 				"cache_salt": "Z3V2bmV3aGxza3ZubGFoZ3Zud3V3ZWZ2bmd0b3V2bnZmc2xpZ3RoZ2x2aQ==",
 			},
-			want: &scheduling.LLMRequestBody{
-				Completions: &scheduling.CompletionsRequest{
+			want: &fwkrh.LLMRequestBody{
+				Completions: &fwkrh.CompletionsRequest{
 					Prompt:    "test prompt",
 					CacheSalt: "Z3V2bmV3aGxza3ZubGFoZ3Zud3V3ZWZ2bmd0b3V2bnZmc2xpZ3RoZ2x2aQ==",
 				},
@@ -470,11 +468,11 @@ func TestOpenAIParser_ParseRequest(t *testing.T) {
 				},
 				"cache_salt": "Z3V2bmV3aGxza3ZubGFoZ3Zud3V3ZWZ2bmd0b3V2bnZmc2xpZ3RoZ2x2aQ==",
 			},
-			want: &scheduling.LLMRequestBody{
-				ChatCompletions: &scheduling.ChatCompletionsRequest{
-					Messages: []scheduling.Message{
-						{Role: "system", Content: scheduling.Content{Raw: "this is a system message"}},
-						{Role: "user", Content: scheduling.Content{Raw: "hello"}},
+			want: &fwkrh.LLMRequestBody{
+				ChatCompletions: &fwkrh.ChatCompletionsRequest{
+					Messages: []fwkrh.Message{
+						{Role: "system", Content: fwkrh.Content{Raw: "this is a system message"}},
+						{Role: "user", Content: fwkrh.Content{Raw: "hello"}},
 					},
 					CacheSalt: "Z3V2bmV3aGxza3ZubGFoZ3Zud3V3ZWZ2bmd0b3V2bnZmc2xpZ3RoZ2x2aQ==",
 				},
@@ -500,8 +498,8 @@ func TestOpenAIParser_ParseRequest(t *testing.T) {
 				"input":        "How do I check if a Python object is an instance of a class?",
 				"instructions": "You are a coding assistant that talks like a pirate.",
 			},
-			want: &scheduling.LLMRequestBody{
-				Responses: &scheduling.ResponsesRequest{
+			want: &fwkrh.LLMRequestBody{
+				Responses: &fwkrh.ResponsesRequest{
 					Input:        "How do I check if a Python object is an instance of a class?",
 					Instructions: "You are a coding assistant that talks like a pirate.",
 				},
@@ -520,8 +518,8 @@ func TestOpenAIParser_ParseRequest(t *testing.T) {
 				"input":      "test input",
 				"cache_salt": "abc123",
 			},
-			want: &scheduling.LLMRequestBody{
-				Responses: &scheduling.ResponsesRequest{
+			want: &fwkrh.LLMRequestBody{
+				Responses: &fwkrh.ResponsesRequest{
 					Input:     "test input",
 					CacheSalt: "abc123",
 				},
@@ -551,9 +549,9 @@ func TestOpenAIParser_ParseRequest(t *testing.T) {
 					{"type": "message", "role": "user", "content": "Hello"},
 				},
 			},
-			want: &scheduling.LLMRequestBody{
-				Conversations: &scheduling.ConversationsRequest{
-					Items: []scheduling.ConversationItem{
+			want: &fwkrh.LLMRequestBody{
+				Conversations: &fwkrh.ConversationsRequest{
+					Items: []fwkrh.ConversationItem{
 						{Type: "message", Role: "user", Content: "Hello"},
 					},
 				},
@@ -572,9 +570,9 @@ func TestOpenAIParser_ParseRequest(t *testing.T) {
 					{"type": "message", "role": "user", "content": "Hello"},
 				},
 			},
-			want: &scheduling.LLMRequestBody{
-				Conversations: &scheduling.ConversationsRequest{
-					Items: []scheduling.ConversationItem{
+			want: &fwkrh.LLMRequestBody{
+				Conversations: &fwkrh.ConversationsRequest{
+					Items: []fwkrh.ConversationItem{
 						{Type: "message", Role: "user", Content: "Hello"},
 					},
 				},
@@ -593,8 +591,8 @@ func TestOpenAIParser_ParseRequest(t *testing.T) {
 				"model":  "gpt-4o",
 				"prompt": "test prompt",
 			},
-			want: &scheduling.LLMRequestBody{
-				Completions: &scheduling.CompletionsRequest{
+			want: &fwkrh.LLMRequestBody{
+				Completions: &fwkrh.CompletionsRequest{
 					Prompt: "test prompt",
 				},
 				ParsedBody: map[string]any{
@@ -647,7 +645,7 @@ func TestOpenAIParser_ParseResponse(t *testing.T) {
 				}
 			}`),
 			want: &fwkrh.ParsedResponse{
-				Usage: &fwkrc.Usage{
+				Usage: &fwkrh.Usage{
 					PromptTokens:     10,
 					CompletionTokens: 20,
 					TotalTokens:      30,
@@ -665,7 +663,7 @@ func TestOpenAIParser_ParseResponse(t *testing.T) {
 				}
 			}`),
 			want: &fwkrh.ParsedResponse{
-				Usage: &fwkrc.Usage{
+				Usage: &fwkrh.Usage{
 					PromptTokens:     15,
 					CompletionTokens: 25,
 					TotalTokens:      40,
@@ -686,11 +684,11 @@ func TestOpenAIParser_ParseResponse(t *testing.T) {
 				}
 			}`),
 			want: &fwkrh.ParsedResponse{
-				Usage: &fwkrc.Usage{
+				Usage: &fwkrh.Usage{
 					PromptTokens:     100,
 					CompletionTokens: 50,
 					TotalTokens:      150,
-					PromptTokenDetails: &fwkrc.PromptTokenDetails{
+					PromptTokenDetails: &fwkrh.PromptTokenDetails{
 						CachedTokens: 40,
 					},
 				},
@@ -707,7 +705,7 @@ func TestOpenAIParser_ParseResponse(t *testing.T) {
 				}
 			}`),
 			want: &fwkrh.ParsedResponse{
-				Usage: &fwkrc.Usage{
+				Usage: &fwkrh.Usage{
 					PromptTokens:     5,
 					CompletionTokens: 5,
 					TotalTokens:      10,
@@ -753,7 +751,7 @@ func TestOpenAIParser_ParseResponse_Streaming(t *testing.T) {
 			name:  "Single data chunk with usage",
 			chunk: []byte("data: {\"usage\":{\"prompt_tokens\":7,\"completion_tokens\":10,\"total_tokens\":17}}\n"),
 			want: &fwkrh.ParsedResponse{
-				Usage: &fwkrc.Usage{
+				Usage: &fwkrh.Usage{
 					PromptTokens:     7,
 					CompletionTokens: 10,
 					TotalTokens:      17,
@@ -764,9 +762,9 @@ func TestOpenAIParser_ParseResponse_Streaming(t *testing.T) {
 			name:  "Usage and DONE in the same multi-line response",
 			chunk: []byte("data: {\"usage\":{\"prompt_tokens\":10,\"prompt_token_details\":{\"cached_tokens\":10}}}\ndata: [DONE]"),
 			want: &fwkrh.ParsedResponse{
-				Usage: &fwkrc.Usage{
+				Usage: &fwkrh.Usage{
 					PromptTokens: 10,
-					PromptTokenDetails: &fwkrc.PromptTokenDetails{
+					PromptTokenDetails: &fwkrh.PromptTokenDetails{
 						CachedTokens: 10,
 					},
 				},
@@ -797,7 +795,7 @@ func TestOpenAIParser_ParseResponse_Streaming(t *testing.T) {
 			name:  "ResponsesAPI streaming with full response",
 			chunk: []byte("event: response.completed\ndata: {\"response\":{\"id\":\"resp_8e38bd02b4f56572\",\"model\":\"Qwen/Qwen3-32B\",\"object\":\"response\",\"usage\":{\"input_tokens\":31,\"input_tokens_details\":{\"cached_tokens\":16},\"output_tokens\":3,\"output_tokens_details\":{\"reasoning_tokens\":0},\"total_tokens\":34}},\"type\":\"response.completed\"}"),
 			want: &fwkrh.ParsedResponse{
-				Usage: &fwkrc.Usage{
+				Usage: &fwkrh.Usage{
 					PromptTokens:     31,
 					CompletionTokens: 3,
 					TotalTokens:      34,
@@ -815,7 +813,7 @@ func TestOpenAIParser_ParseResponse_Streaming(t *testing.T) {
 			name:  "ResponsesAPI with multiple events extracts from completed",
 			chunk: []byte("event: response.output_text.delta\ndata: {\"delta\":\"Hello\",\"type\":\"response.output_text.delta\"}\n\nevent: response.completed\ndata: {\"response\":{\"usage\":{\"input_tokens\":39,\"output_tokens\":10,\"total_tokens\":49}},\"type\":\"response.completed\"}"),
 			want: &fwkrh.ParsedResponse{
-				Usage: &fwkrc.Usage{
+				Usage: &fwkrh.Usage{
 					PromptTokens:     39,
 					CompletionTokens: 10,
 					TotalTokens:      49,
