+++ pkg/epp/saturationdetector/framework/plugins/concurrencydetector/token_estimator_test.go
@@ -22,6 +22,7 @@ import (
 	"github.com/stretchr/testify/require"
 
 	framework "sigs.k8s.io/gateway-api-inference-extension/pkg/epp/framework/interface/scheduling"
+	fwkrh "sigs.k8s.io/gateway-api-inference-extension/pkg/epp/framework/interface/requesthandling"
 )
 
 func TestSimpleTokenEstimator_Estimate(t *testing.T) {
@@ -52,8 +53,8 @@ func TestSimpleTokenEstimator_Estimate(t *testing.T) {
 		{
 			name: "Less than 4 characters",
 			request: &framework.LLMRequest{
-				Body: &framework.LLMRequestBody{
-					Completions: &framework.CompletionsRequest{
+				Body: &fwkrh.LLMRequestBody{
+					Completions: &fwkrh.CompletionsRequest{
 						Prompt: "123",
 					},
 				},
@@ -63,8 +64,8 @@ func TestSimpleTokenEstimator_Estimate(t *testing.T) {
 		{
 			name: "Completions Request",
 			request: &framework.LLMRequest{
-				Body: &framework.LLMRequestBody{
-					Completions: &framework.CompletionsRequest{
+				Body: &fwkrh.LLMRequestBody{
+					Completions: &fwkrh.CompletionsRequest{
 						Prompt: "Hello, world!",
 					},
 				},
@@ -74,8 +75,8 @@ func TestSimpleTokenEstimator_Estimate(t *testing.T) {
 		{
 			name: "Completions with empty prompt",
 			request: &framework.LLMRequest{
-				Body: &framework.LLMRequestBody{
-					Completions: &framework.CompletionsRequest{
+				Body: &fwkrh.LLMRequestBody{
+					Completions: &fwkrh.CompletionsRequest{
 						Prompt: "",
 					},
 				},
@@ -85,8 +86,8 @@ func TestSimpleTokenEstimator_Estimate(t *testing.T) {
 		{
 			name: "Completions with exactly 4 characters",
 			request: &framework.LLMRequest{
-				Body: &framework.LLMRequestBody{
-					Completions: &framework.CompletionsRequest{
+				Body: &fwkrh.LLMRequestBody{
+					Completions: &fwkrh.CompletionsRequest{
 						Prompt: "1234",
 					},
 				},
@@ -96,13 +97,13 @@ func TestSimpleTokenEstimator_Estimate(t *testing.T) {
 		{
 			name: "Chat Completions Request with Structured content",
 			request: &framework.LLMRequest{
-				Body: &framework.LLMRequestBody{
-					ChatCompletions: &framework.ChatCompletionsRequest{
-						Messages: []framework.Message{
+				Body: &fwkrh.LLMRequestBody{
+					ChatCompletions: &fwkrh.ChatCompletionsRequest{
+						Messages: []fwkrh.Message{
 							{
 								Role: "user",
-								Content: framework.Content{
-									Structured: []framework.ContentBlock{
+								Content: fwkrh.Content{
+									Structured: []fwkrh.ContentBlock{
 										{
 											Type: "text",
 											Text: "This is a longer message.",
@@ -119,12 +120,12 @@ func TestSimpleTokenEstimator_Estimate(t *testing.T) {
 		{
 			name: "Chat Completions with Raw content",
 			request: &framework.LLMRequest{
-				Body: &framework.LLMRequestBody{
-					ChatCompletions: &framework.ChatCompletionsRequest{
-						Messages: []framework.Message{
+				Body: &fwkrh.LLMRequestBody{
+					ChatCompletions: &fwkrh.ChatCompletionsRequest{
+						Messages: []fwkrh.Message{
 							{
 								Role: "user",
-								Content: framework.Content{
+								Content: fwkrh.Content{
 									Raw: "This is raw content.",
 								},
 							},
@@ -137,21 +138,21 @@ func TestSimpleTokenEstimator_Estimate(t *testing.T) {
 		{
 			name: "Chat Completions with multiple messages",
 			request: &framework.LLMRequest{
-				Body: &framework.LLMRequestBody{
-					ChatCompletions: &framework.ChatCompletionsRequest{
-						Messages: []framework.Message{
+				Body: &fwkrh.LLMRequestBody{
+					ChatCompletions: &fwkrh.ChatCompletionsRequest{
+						Messages: []fwkrh.Message{
 							{
 								Role: "user",
-								Content: framework.Content{
-									Structured: []framework.ContentBlock{
+								Content: fwkrh.Content{
+									Structured: []fwkrh.ContentBlock{
 										{Type: "text", Text: "Hi"},
 									},
 								},
 							},
 							{
 								Role: "assistant",
-								Content: framework.Content{
-									Structured: []framework.ContentBlock{
+								Content: fwkrh.Content{
+									Structured: []fwkrh.ContentBlock{
 										{Type: "text", Text: "Hello"},
 									},
 								},
@@ -165,9 +166,9 @@ func TestSimpleTokenEstimator_Estimate(t *testing.T) {
 		{
 			name: "Chat Completions with empty messages",
 			request: &framework.LLMRequest{
-				Body: &framework.LLMRequestBody{
-					ChatCompletions: &framework.ChatCompletionsRequest{
-						Messages: []framework.Message{},
+				Body: &fwkrh.LLMRequestBody{
+					ChatCompletions: &fwkrh.ChatCompletionsRequest{
+						Messages: []fwkrh.Message{},
 					},
 				},
 			},
@@ -197,8 +198,8 @@ func TestSimpleTokenEstimator_Estimate_CustomConfig(t *testing.T) {
 		{
 			name: "Empty prompt with custom config",
 			request: &framework.LLMRequest{
-				Body: &framework.LLMRequestBody{
-					Completions: &framework.CompletionsRequest{
+				Body: &fwkrh.LLMRequestBody{
+					Completions: &fwkrh.CompletionsRequest{
 						Prompt: "",
 					},
 				},
@@ -208,8 +209,8 @@ func TestSimpleTokenEstimator_Estimate_CustomConfig(t *testing.T) {
 		{
 			name: "4 chars with custom config",
 			request: &framework.LLMRequest{
-				Body: &framework.LLMRequestBody{
-					Completions: &framework.CompletionsRequest{
+				Body: &fwkrh.LLMRequestBody{
+					Completions: &fwkrh.CompletionsRequest{
 						Prompt: "1234",
 					},
 				},
@@ -219,8 +220,8 @@ func TestSimpleTokenEstimator_Estimate_CustomConfig(t *testing.T) {
 		{
 			name: "More than 4 chars with custom config",
 			request: &framework.LLMRequest{
-				Body: &framework.LLMRequestBody{
-					Completions: &framework.CompletionsRequest{
+				Body: &fwkrh.LLMRequestBody{
+					Completions: &fwkrh.CompletionsRequest{
 						Prompt: "This is a longer message.",
 					},
 				},
